+++ packages/gamut/src/typings/responsive-properties.ts
@@ -1,0 +1,17 @@
+export type ResponsiveProperty<T> =
+  | T
+  | {
+      xs: T;
+      sm?: T;
+      md?: T;
+      lg?: T;
+    };
+
+export type OptionalResponsiveProperty<T> =
+  | T
+  | {
+      xs?: T;
+      sm?: T;
+      md?: T;
+      lg?: T;
+    };